refactor: move Variabale FinalForm into separate component
