+++ operate/client/src/App/ProcessInstance/BottomPanel/VariablePanel/VariablesContent.tsx
@@ -7,26 +7,18 @@
  */
 
 import {observer} from 'mobx-react';
-import {Form as ReactFinalForm} from 'react-final-form';
-import {type VariableFormValues} from 'modules/types/variables';
 
 import {Content, EmptyMessageContainer} from './styled';
-import arrayMutators from 'final-form-arrays';
 import {ErrorMessage} from 'modules/components/ErrorMessage';
 import {EmptyMessage} from 'modules/components/EmptyMessage';
 import {Loading} from '@carbon/react';
-import {VariablesForm} from './VariablesForm';
-import {notificationsStore} from 'modules/stores/notifications';
-import {useProcessInstancePageParams} from 'App/ProcessInstance/useProcessInstancePageParams';
-import {addVariable, getScopeId, updateVariable} from 'modules/utils/variables';
-import {useQueryClient} from '@tanstack/react-query';
+import {getScopeId} from 'modules/utils/variables';
 import {useVariables} from 'modules/queries/variables/useVariables';
-import {queryKeys} from 'modules/queries/queryKeys';
+import {VariablesFinalForm} from './VariablesFinalForm';
 
 const VariablesContent: React.FC = observer(() => {
-  const {processInstanceId = ''} = useProcessInstancePageParams();
-  const queryClient = useQueryClient();
   const {displayStatus} = useVariables();
+  const scopeId = getScopeId();
 
   if (displayStatus === 'error') {
     return (
@@ -52,68 +44,7 @@ const VariablesContent: React.FC = observer(() => {
       {displayStatus === 'spinner' && (
         <Loading data-testid="variables-spinner" />
       )}
-      <ReactFinalForm<VariableFormValues>
-        mutators={{
-          ...arrayMutators,
-          triggerValidation(fieldsToValidate: string[], state, {changeValue}) {
-            fieldsToValidate.forEach((fieldName) => {
-              changeValue(state, fieldName, (n) => n);
-            });
-          },
-        }}
-        key={getScopeId()}
-        render={(props) => <VariablesForm {...props} />}
-        onSubmit={async (values, form) => {
-          const {initialValues} = form.getState();
-
-          const {name, value} = values;
-
-          if (name === undefined || value === undefined) {
-            return;
-          }
-
-          const params = {
-            id: processInstanceId,
-            name,
-            value,
-            invalidateQueries: () => {
-              queryClient.invalidateQueries({
-                queryKey: queryKeys.variables.search(),
-              });
-            },
-            onSuccess: () => {
-              notificationsStore.displayNotification({
-                kind: 'success',
-                title: 'Variable added',
-                isDismissable: true,
-              });
-
-              form.reset({});
-            },
-            onError: (statusCode: number) => {
-              notificationsStore.displayNotification({
-                kind: 'error',
-                title: 'Variable could not be saved',
-                subtitle:
-                  statusCode === 403 ? 'You do not have permission' : undefined,
-                isDismissable: true,
-              });
-
-              form.reset({});
-            },
-          };
-
-          if (initialValues.name === '') {
-            const result = await addVariable(params);
-            if (result === 'VALIDATION_ERROR') {
-              return {name: 'Name should be unique'};
-            }
-          } else if (initialValues.name === name) {
-            updateVariable(params);
-            form.reset({});
-          }
-        }}
-      />
+      {scopeId !== null && <VariablesFinalForm scopeId={scopeId} />}
     </Content>
   );
 });
